New mapper type
Extract mapping logic to a new type on it's own file.
