+++ FeedAPIChallenge/Feed API/RemoteFeedLoader.swift
@@ -22,8 +22,8 @@ public final class RemoteFeedLoader: FeedLoader {
 		client.get(from: url) { result in
 			switch result {
 			case let .success((data, response)):
-				if response.statusCode == 200, let responseData = try? JSONDecoder().decode(FeedImagesResponse.self, from: data) {
-					completion(.success(responseData.items.map { $0.image }))
+				if let items = try? FeedImageMapper.map(data, response) {
+					completion(.success(items))
 				} else {
 					completion(.failure(Error.invalidData))
 				}
@@ -31,20 +31,5 @@ public final class RemoteFeedLoader: FeedLoader {
 				completion(.failure(Error.connectivity))
 			}
 		}
-	}
-}
-
-private struct FeedImagesResponse: Decodable {
-	let items: [Image]
-}
-
-private struct Image: Decodable {
-	let image_id: UUID
-	let image_desc: String?
-	let image_loc: String?
-	let image_url: URL
-
-	var image: FeedImage {
-		return FeedImage(id: image_id, description: image_desc, location: image_loc, url: image_url)
 	}
 }
